[Finishes #109609786] Refactoring and Review
Did refactoring of the serializers and models with @Eric.
So from now henceforth, all the serializers will be in one
model app/serializers.py and the models are grouped in various
modules in the models package

For the migrations to access the model classes, we did
imports in models/__init__.py -- also for the
settings.AUTH_USER_MODEL
