+++ app/models/__init__.py
@@ -1,0 +1,6 @@
+
+# These imports are made to be used by the migration script
+# and in settings.AUTH_USER_MODEL
+from user import User, UserAuthentication, Member
+from project import Project, Team
+from story import Story, Task
