+++ app/templates/user/profile.html
@@ -1,0 +1,12 @@
+{% extends 'partials/layout_basic.html' %}
+
+{% load staticfiles %}
+
+{% block content %}
+	
+	<h4>This is the profile page</h4>
+	<div class="mdl-card__supporting-text">
+		<h6>{{ message }}</h6>
+	</div>
+
+{% endblock %}
