+++ app/templates/user/register.html
@@ -1,0 +1,46 @@
+{% extends 'partials/layout_basic.html' %} {% load staticfiles %} {% block content %}
+<div class="row">
+    <div class="col s12 m6">
+        <div class="card centered">
+            <div class="card-action light-blue darken-2">
+                <div class="card-content">
+                    <span class="card-title activator white-text text-darken-4">Register</span>
+                </div>
+            </div>
+
+            <div class="card-content cyan-text">
+      				<div class="card-text">
+                {{ message }}
+                <form action="#" method="POST">
+      					{% csrf_token %}
+      					{{ form.non_field_errors }}
+        					<div class="input-field">
+        					<i class="material-icons prefix mdl-color-text--white-400">account_circle</i>
+        						{{ form.username.errors }}
+        						<label class="" for="{{ form.username.id_for_label }}">Username</label>
+        						{{ form.username }}
+        					</div>
+        					<div class="input-field">
+        					<i class="material-icons prefix">email</i>
+
+        						{{ form.email.errors }}
+        						<label class="" for="{{ form.email.id_for_label }}">Email Address</label>
+        						{{ form.email }}
+        					</div>
+        					<div class="input-field">
+        					<i class="material-icons prefix">vpn_key</i>
+        						{{ form.password.errors }}
+        						<label class="" for="{{ form.password.id_for_label }}">Password</label>
+        						{{ form.password }}					
+        					</div>
+      				</div>
+              <div class="card-action">
+                <input class="btn blue darker-2" type="submit" value="Register" />
+                <a href="{% url 'login' %}">Login</a>
+              </div>
+              </form>
+          </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
